split compute jobs into two different files
