+++ tests/test_compute_jobs.py
@@ -2,12 +2,13 @@
 import pytest
 import requests_mock
 
-from c20_client.compute_jobs import (
-    reggov_api_doc_error,
-    compute_jobs,
-    get_number_of_docs,
-    get_response_from_api
+from c20_client.reggov_api_doc_error import (
+    IncorrectApiKeyException,
+    ExceedCallLimitException
 )
+
+from c20_client.compute_jobs import compute_jobs, get_number_of_docs
+from c20_client.get_documents import get_response_from_api
 
 URL = "https://api.data.gov:443/regulations/v3/documents.json?api_key="
 API_KEY = "Valid"
@@ -31,7 +32,8 @@ def test_valid_jobs_list():
         mock.get(URL + API_KEY,
                  json={"result": "The test is successful",
                        "totalNumRecords": 2000})
-        job_list = compute_jobs(API_KEY, START_DATE, END_DATE)
+        data = get_response_from_api(API_KEY, START_DATE, END_DATE)
+        job_list = compute_jobs(data)
 
         assert job_list == [0, 1000]
 
@@ -52,7 +54,7 @@ def test_bad_api_key():
         mock.get(URL + 'INVALID',
                  json='The test yields a bad api key', status_code=403)
 
-        with pytest.raises(reggov_api_doc_error.IncorrectApiKeyException):
+        with pytest.raises(IncorrectApiKeyException):
             get_response_from_api('INVALID', START_DATE, END_DATE)
 
 
@@ -61,5 +63,5 @@ def test_overused_api_key():
         mock.get(URL + API_KEY,
                  json='The test yields a overused api key', status_code=429)
 
-        with pytest.raises(reggov_api_doc_error.ExceedCallLimitException):
+        with pytest.raises(ExceedCallLimitException):
             get_response_from_api(API_KEY, START_DATE, END_DATE)
